adding the application to collect the metrics and saving the artifact to json and html and radar viz
